+++ src/components/adver/categories/CategoryCard/CategoryCard.module.scss
