+++ src/components/adver/categories/CategoryCard/CategoryCard.tsx
