+++ src/components/adver/categories/index.ts
@@ -1,0 +1,3 @@
+export { CategoryTreeModal } from "./AdverSubCategoriesDesktop/CategoryTreeModal";
+export { CategoryCard } from "./CategoryCard/CategoryCard";
+export { CategorySelectModal } from "./CategorySelectModal/CategorySelectModal"
