+++ src/components/adver/form/index.ts
@@ -1,0 +1,4 @@
+// export { default as PriceBlock } from "./PriceBlock/PriceBlock";
+export { default as AdverDesktop } from "./AdverDesktop/AdverDesktop";
+export { default as AdverMobile } from "./AdverMobile/AdverMobile";
+export { default as CityAutocomplete } from "./LocationBlock/CityAutocomplete/CityAutocomplete";
