+++ src/components/adver/index.ts
@@ -1,0 +1,4 @@
+export * from "./form";
+export * from "./preview";
+export * from "./photos";
+export * from "./categories";
