+++ src/components/adver/photos/index.ts
@@ -1,0 +1,1 @@
+export { default as AdverPhotoList } from "./AdverPhotoList/AdverPhotoList";
