+++ src/components/adver/preview/AdverPreviewModal/AdverPreviewModal.module.scss
