+++ src/components/adver/preview/AdverPreviewSlider/AdverPreviewSlider.tsx
