+++ src/components/adver/preview/DescriptionPreview/DescriptionPreview.module.scss
