+++ src/components/adver/preview/LocationPreview/LocationPreview.module.scss
