+++ src/components/adver/preview/PreviewActions/PreviewActions.module.scss
