+++ src/components/adver/preview/PreviewHeader/PreviewHeader.module.scss
