+++ src/components/adver/preview/PreviewPrice/PreviewPrice.module.scss
