+++ src/components/adver/preview/PreviewPrice/PreviewPrice.tsx
