+++ src/components/adver/preview/SellerPreview/SellerPreview.module.scss
