+++ src/components/adver/preview/SellerPreview/SellerPreview.tsx
