+++ src/components/adver/preview/index.ts
@@ -1,0 +1,7 @@
+export { default as PreviewHeader } from "./PreviewHeader/PreviewHeader";
+export { default as PreviewPrice } from "./PreviewPrice/PreviewPrice";
+export { default as SellerPreview } from "./SellerPreview/SellerPreview";
+export { default as LocationPreview } from "./LocationPreview/LocationPreview";
+export { default as DeliveryMethodsPreview } from "./DeliveryMethodsPreview/DeliveryMethodsPreview";
+export { default as DescriptionPreview } from "./DescriptionPreview/DescriptionPreview";
+export { default as PreviewActions } from "./PreviewActions/PreviewActions";
